Add `pre_insert_txs` flag
