+++ core/lib/zksync_core/src/state_keeper/io/tests/mod.rs
@@ -246,6 +246,7 @@ async fn processing_storage_logs_when_sealing_miniblock() {
         protocol_version: Some(ProtocolVersionId::latest()),
         l2_erc20_bridge_addr: Address::default(),
         consensus: None,
+        pre_insert_txs: false,
     };
     let mut conn = connection_pool
         .access_storage_tagged("state_keeper")
@@ -323,6 +324,7 @@ async fn processing_events_when_sealing_miniblock() {
         protocol_version: Some(ProtocolVersionId::latest()),
         l2_erc20_bridge_addr: Address::default(),
         consensus: None,
+        pre_insert_txs: false,
     };
     let mut conn = pool.access_storage_tagged("state_keeper").await.unwrap();
     conn.protocol_versions_dal()
@@ -437,6 +439,7 @@ async fn miniblock_sealer_handle_blocking() {
         MiniblockNumber(1),
         Address::default(),
         None,
+        false,
     );
     sealer_handle.submit(seal_command).await;
 
@@ -446,6 +449,7 @@ async fn miniblock_sealer_handle_blocking() {
         MiniblockNumber(2),
         Address::default(),
         None,
+        false,
     );
     {
         let submit_future = sealer_handle.submit(seal_command);
@@ -475,6 +479,7 @@ async fn miniblock_sealer_handle_blocking() {
         MiniblockNumber(3),
         Address::default(),
         None,
+        false,
     );
     sealer_handle.submit(seal_command).await;
     let command = sealer.commands_receiver.recv().await.unwrap();
@@ -495,6 +500,7 @@ async fn miniblock_sealer_handle_parallel_processing() {
             MiniblockNumber(i),
             Address::default(),
             None,
+            false,
         );
         sealer_handle.submit(seal_command).await;
     }
