+++ core/lib/zksync_core/src/state_keeper/metrics.rs
@@ -168,7 +168,6 @@ pub(super) enum L1BatchSealStage {
     FilterWrittenSlots,
     InsertInitialWrites,
     CommitL1Batch,
-    ExternalNodeStoreTransactions,
 }
 
 /// Buckets for positive integer, not-so-large values (e.g., initial writes count).
@@ -221,10 +220,6 @@ impl L1BatchMetrics {
             latency_per_unit: &self.sealed_entity_per_unit[&stage],
         }
     }
-
-    pub(crate) fn start_storing_on_en(&self) -> LatencyObserver<'_> {
-        self.sealed_time_stage[&L1BatchSealStage::ExternalNodeStoreTransactions].start()
-    }
 }
 
 #[vise::register]
@@ -241,6 +236,7 @@ pub(super) enum MiniblockQueueStage {
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, EncodeLabelValue)]
 #[metrics(rename_all = "snake_case")]
 pub(super) enum MiniblockSealStage {
+    PreInsertTxs,
     InsertMiniblockHeader,
     MarkTransactionsInMiniblock,
     InsertStorageLogs,
